Added the complete Activity List for a LAN

+++ resources/views/lan/show_parts/activities_card.blade.php
@@ -9,7 +9,7 @@
 									@if ($userIsLanAdmin)
 									<a class="btn btn-success shadow-sm float-right ml-2" href="{{ route('activity.create', $lan->id) }}"><i class='fa fa-plus'></i></a>
 									@endif
-									<a class="btn btn-outline-primary shadow-sm float-right" href="{{ route('lan.add_game', $lan->id) }}"><i class='fa fa-list'></i> All</a>
+									<a class="btn btn-outline-primary shadow-sm float-right" href="{{ route('lan.activity_list', $lan->id) }}"><i class='fa fa-list'></i> All</a>
 								</div>
 							</div>
 						</div>
@@ -38,12 +38,15 @@
 													{!! Form::open(['onsubmit'=>'return false;']) !!}
 														<button class="btn btn-success mr-2" id="activity-view-{{$activity->id}}" onclick="openActivity({{$activity->id}})"><i class='fa fa-eye'></i> View</button>
 													{{ Form::close() }}
+@if ($userIsLanAdmin)
+
 													{!! Form::open(['method'=>'get','url'=>route('activity.edit', array('lan' => $lan->id, 'activity' => $activity->id))]) !!}
 														<button class="btn btn-warning mr-2"><i class='fa fa-edit'></i> Edit</button>
 													{{ Form::close() }}
 													{!! Form::open(['method' => 'delete','url'=>'', 'onsubmit'=>'return removeActivity(event, '.$lan->id.', '.$activity->id.')']) !!}
 														<button type="submit" class="btn btn-danger"><i class='fa fa-trash'></i> Delete</button>
 													{!! Form::close() !!}
+@endif
 												</div>
 											</td>
 										</tr>
